pull request 20240306

+++ setup.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-#-*- coding:utf-8 -*-
+# -*- coding:utf-8 -*-
 
 from setuptools import setup, find_packages
 import sys
@@ -10,9 +10,9 @@ import platform
 import click
 
 
-#------------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
 # 前置检查
-#------------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
 def check_xmake():
     """检查是否按照了编译工具 xmake"""
     print("checking xmake ...")
@@ -25,7 +25,7 @@ def get_python_version():
     py_version = platform.python_version_tuple()
     min_version = int(py_version[1])
     main_version = int(py_version[0])
-    #py_version = main_version * 10 + min_version if min_version < 10 else main_version * 100 + min_version
+    # py_version = main_version * 10 + min_version if min_version < 10 else main_version * 100 + min_version
     py_version = f"{main_version}.{min_version}"
     print(f'current python version: {py_version}')
     return py_version
@@ -102,10 +102,10 @@ def clear_with_python_changed(mode):
         shutil.rmtree(build_pywrap_dir)
 
 
-#------------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
 # 执行构建
-#------------------------------------------------------------------------------
-def start_build(verbose=False, mode='release', feedback=True, worker_num=2):
+# ------------------------------------------------------------------------------
+def start_build(verbose=False, mode='release', feedback=True, worker_num=2, low_precision=False):
     """ 执行编译 """
     global g_verbose
     g_verbose = verbose
@@ -118,13 +118,13 @@ def start_build(verbose=False, mode='release', feedback=True, worker_num=2):
 
     py_version = current_compile_info['py_version']
 
-    #如果 python版本或者编译模式发生变化，则编译依赖的 boost 库（boost.python)
+    # 如果 python版本或者编译模式发生变化，则重新编译
     history_compile_info = get_history_compile_info()
     if py_version != history_compile_info[
             'py_version'] or history_compile_info['mode'] != mode:
         clear_with_python_changed(mode)
-        cmd = "xmake f {} -c -y -m {} --pyver={} --feedback={}".format(
-            "-v -D" if verbose else "", mode, py_version, feedback)
+        cmd = "xmake f {} -c -y -m {} --feedback={} -k {} --low_precision={}".format(
+            "-v -D" if verbose else "", mode, feedback, "shared" if mode == 'release' else "static", low_precision)
         print(cmd)
         os.system(cmd)
 
@@ -143,9 +143,9 @@ def start_build(verbose=False, mode='release', feedback=True, worker_num=2):
     save_current_compile_info(current_compile_info)
 
 
-#------------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
 # 控制台命令
-#------------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
 
 
 @click.group()
@@ -169,9 +169,14 @@ def cli():
                   'lsan'
               ]),
               help='编译模式')
-def build(verbose, mode, feedback, j):
+@click.option('-low_precision',
+              '--low_precision',
+              default=False,
+              type=bool,
+              help='使用低精度版本')
+def build(verbose, mode, feedback, j, low_precision):
     """ 执行编译 """
-    start_build(verbose, mode, feedback, j)
+    start_build(verbose, mode, feedback, j, low_precision)
 
 
 @click.command()
@@ -193,13 +198,18 @@ def build(verbose, mode, feedback, j):
               ]),
               help='编译模式')
 @click.option('-case', '--case', default='', help="执行指定的 TestCase")
-def test(all, compile, verbose, mode, case, feedback, j):
+@click.option('-low_precision',
+              '--low_precision',
+              default=False,
+              type=bool,
+              help='使用低精度版本')
+def test(all, compile, verbose, mode, case, feedback, j, low_precision):
     """ 执行单元测试 """
     current_compile_info = get_current_compile_info()
     current_compile_info['mode'] = mode
     history_compile_info = get_history_compile_info()
     if compile or current_compile_info != history_compile_info:
-        start_build(verbose, mode, feedback, j)
+        start_build(verbose, mode, feedback, j, low_precision)
     if all:
         os.system("xmake -j {} -b {} unit-test".format(
             j, "-v -D" if verbose else ""))
@@ -216,7 +226,9 @@ def clear_build():
     """ 清除当前编译设置及结果 """
     if os.path.lexists('.xmake'):
         print('delete .xmake')
-        shutil.rmtree('.xmake')
+        shutil.rmtree('.xmake', True)
+        if sys.platform == 'win32':
+            os.system("rmdir .xmake /s /q")
     if os.path.lexists('build'):
         print('delete build')
         shutil.rmtree('build')
@@ -231,7 +243,6 @@ def clear_build():
         if file not in ("__init__.py", "__pycache__"):
             os.remove(f'hikyuu/cpp/{file}')
     print('clear finished!')
-    os.system("xmake clean")
 
 
 @click.command()
@@ -258,22 +269,62 @@ def uninstall():
     print("Uninstall finished!")
 
 
+def copy_include(install_dir):
+    src_path = 'hikyuu_cpp/hikyuu'
+    dst_path = f'{install_dir}/include'
+
+    for root, dirs, files in os.walk(src_path):
+        for p in dirs:
+            dst_p = f'{dst_path}/{root[11:]}/{p}'
+            if not os.path.lexists(dst_p):
+                os.makedirs(dst_p)
+            shutil.copy('hikyuu/cpp/__init__.py', dst_p)
+
+        for fname in files:
+            if len(fname) > 2 and fname[-2:] == ".h":
+                dst_p = f'{dst_path}/{root[11:]}'
+                if not os.path.lexists(dst_p):
+                    os.makedirs(dst_p)
+                shutil.copy(f'{root}/{fname}', dst_p)
+
+    dst_path = f'{install_dir}/include/hikyuu/python'
+    if not os.path.lexists(dst_path):
+        os.makedirs(dst_path)
+    shutil.copy('hikyuu_pywrap/pybind_utils.h', dst_path)
+    shutil.copy('hikyuu_pywrap/pickle_support.h', dst_path)
+    shutil.copy('hikyuu/cpp/__init__.py', dst_path)
+    shutil.copy('hikyuu/cpp/__init__.py', f'{install_dir}/include')
+    shutil.copy('hikyuu/cpp/__init__.py', f'{install_dir}/include/hikyuu')
+
+
 @click.command()
-def install():
+@click.option('-j', '--j', default=2, help="并行编译数量")
+@click.option('-o', '--o', help="指定的安装目录")
+@click.option('-low_precision',
+              '--low_precision',
+              default=False,
+              type=bool,
+              help='使用低精度版本')
+def install(j, o, low_precision):
     """ 编译并安装 Hikyuu python 库 """
-    start_build(False, 'release')
-    if sys.platform == 'win32':
-        install_dir = sys.base_prefix + "\\Lib\\site-packages\\hikyuu"
-    else:
-        usr_dir = os.path.expanduser('~')
-        install_dir = '{}/.local/lib/python{}/site-packages/hikyuu'.format(
-            usr_dir, get_python_version())
-        try:
-            shutil.rmtree(install_dir)
-        except:
-            pass
-        os.makedirs(install_dir)
-    os.system('xmake install -o "{}"'.format(install_dir))
+    install_dir = o
+    if install_dir is None:
+        if sys.platform == 'win32':
+            install_dir = sys.base_prefix + "\\Lib\\site-packages\\hikyuu"
+        else:
+            usr_dir = os.path.expanduser('~')
+            install_dir = '{}/.local/lib/python{}/site-packages/hikyuu'.format(
+                usr_dir, get_python_version())
+            try:
+                shutil.rmtree(install_dir)
+            except:
+                pass
+
+    start_build(False, 'release', True, j, low_precision)
+
+    shutil.copytree("./hikyuu", install_dir)
+
+    copy_include(install_dir)
 
 
 @click.command()
@@ -283,26 +334,34 @@ def install():
               default=True,
               type=bool,
               help='允许发送反馈信息')
-def wheel(feedback, j):
+@click.option('-low_precision',
+              '--low_precision',
+              default=False,
+              type=bool,
+              help='使用低精度版本')
+def wheel(feedback, j, low_precision):
     """ 生成 python 的 wheel 安装包 """
     # 清理之前遗留的打包产物
     clear_build()
 
     # 尝试编译
-    start_build(False, 'release', feedback, j)
+    start_build(False, 'release', feedback, j, low_precision)
+
+    copy_include('hikyuu')
 
     # 构建打包命令
     print("start pacakaging bdist_wheel ...")
     current_plat = sys.platform
+    cpu_arch = platform.machine()
     current_bits = 64 if sys.maxsize > 2**32 else 32
     if current_plat == 'win32' and current_bits == 64:
-        plat = "win-amd64"
+        plat = "win_amd64"
     elif current_plat == 'win32' and current_bits == 32:
         plat = "win32"
     elif current_plat == 'linux' and current_bits == 64:
-        plat = "manylinux1_x86_64"
+        plat = f"manylinux1_{cpu_arch}"
     elif current_plat == 'linux' and current_bits == 32:
-        plat = "manylinux1_i386"
+        plat = f"manylinux1_{cpu_arch}"
     elif current_plat == 'darwin' and current_bits == 32:
         plat = "macosx_i686"
     elif current_plat == 'darwin' and current_bits == 64:
@@ -310,6 +369,9 @@ def wheel(feedback, j):
     else:
         print("*********尚未实现该平台的支持*******")
         return
+
+    if low_precision:
+        plat = f"{plat}_low_precision"
 
     py_version = get_python_version()
     main_ver, min_ver = py_version.split('.')
@@ -323,6 +385,8 @@ def wheel(feedback, j):
             main_ver, min_ver, plat)
         print(cmd)
         os.system(cmd)
+
+    shutil.rmtree('hikyuu/include', True)
 
 
 @click.command()
@@ -340,9 +404,9 @@ def upload():
         os.system("twine upload dist/*")
 
 
-#------------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
 # 添加 click 命令
-#------------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
 cli.add_command(build)
 cli.add_command(test)
 cli.add_command(clear)
